Finishing the menu tab, and removing the settings screen.
